All changes made during debugging session on 11/15.

+++ simulator/src/simulator/L1Cache.java
@@ -2,10 +2,10 @@ package simulator;
 
 public class L1Cache {
 
-    int p, b, n, a;
+    long p, b, n, a;
     L1CacheEntry[][] cache;
 
-    public L1Cache(int p, int b, int n, int a) {
+    public L1Cache(long p, long b, long n, long a) {
         this.p = p;
         this.b = b;
         this.n = n;
@@ -14,7 +14,7 @@ public class L1Cache {
         cache = new L1CacheEntry[(int) Math.pow(2, a)][(int) Math.pow(2, n - a - b)];
     }
     
-    public boolean hit(int address, int cycle, boolean read) {
+    public boolean hit(long address, long cycle, boolean read) {
         Block.MSIState state = getState(address);
         if((read && (state == Block.MSIState.SHARED)) || state == Block.MSIState.MODIFIED) {
             L1CacheEntry entry = getEntry(address);
@@ -25,7 +25,7 @@ public class L1Cache {
         }
     }
     
-    public Block.MSIState getState(int address) {
+    public Block.MSIState getState(long address) {
         L1CacheEntry entry = getEntry(address);
         if(entry == null) {
             return Block.MSIState.INVALID;
@@ -34,17 +34,17 @@ public class L1Cache {
         }
     }
     
-    public int setState(int address, Block.MSIState state, int cycle, boolean own) {
+    public long setState(long address, Block.MSIState state, long cycle, boolean own) {
         L1CacheEntry entry = getEntry(address);
         if(!own) {
             entry.setState(state);
         } else if(entry == null && state != Block.MSIState.INVALID) { // Block is not in the L1 cache
-            int tag = Block.L1tag(address, p, b, n, a);
+            long tag = Block.L1tag(address, p, b, n, a);
             L1CacheEntry newEntry = new L1CacheEntry(tag, state, address);
             newEntry.touch(cycle);
             
-            int cacheIndex = Block.L1cacheIndex(address, b, n, a);
-            int minLastCycle = Integer.MAX_VALUE;
+            int cacheIndex = (int) Block.L1cacheIndex(address, b, n, a);
+            long minLastCycle = Integer.MAX_VALUE;
             int LRUway = 0;
             
             // Search the ways for empty slot, else evict
@@ -61,7 +61,7 @@ public class L1Cache {
             }
             
             // Set was full, evicting the LRU block
-            int evictAddress = cache[LRUway][cacheIndex].getAddress();
+            long evictAddress = cache[LRUway][cacheIndex].getAddress();
             cache[LRUway][cacheIndex] = newEntry;
             return evictAddress;
         } else { // Block is in the L1 cache, just change the state
@@ -72,12 +72,12 @@ public class L1Cache {
         return -1;
     }
     
-    public L1CacheEntry getEntry(int address) {
-        int cacheIndex = Block.L1cacheIndex(address, b, n, a);
+    public L1CacheEntry getEntry(long address) {
+        int cacheIndex = (int) Block.L1cacheIndex(address, b, n, a);
         for (int way = 0; way < (int) Math.pow(2, a); way++) {
             if (cache[way][cacheIndex] != null) {
                 L1CacheEntry entry = cache[way][cacheIndex];
-                int tag = Block.L1tag(address, p, b, n, a);
+                long tag = Block.L1tag(address, p, b, n, a);
                 if (entry.getTag() == tag && entry.getState() != Block.MSIState.INVALID) { // Block is in cache, return it
                     return entry;
                 }
